+++ src/Ampersand/FSpec/ToFSpec/ADL2FSpec.hs
@@ -338,7 +338,8 @@ makeFSpec opts context
             -- Each interface gets all attributes that are required to create and delete the object.
             -- All total attributes must be included, because the interface must allow an object to be deleted.
         in
-        [Ifc { ifcname     = name c
+        [Ifc { ifcIsAPI    = False
+             , ifcname     = name c
              , ifcObj      = ObjectDef
                                  { objnm   = name c
                                  , objpos  = Origin "generated object: step 4a - default theme"
@@ -364,7 +365,8 @@ makeFSpec opts context
      --end otherwise: default theme
      --end stap4a
      step4b --generate lists of concept instances for those concepts that have a generated INTERFACE in step4a
-      = [Ifc { ifcname     = nm
+      = [Ifc { ifcIsAPI    = False
+             , ifcname     = nm
              , ifcObj      = ObjectDef
                                  { objnm   = nm
                                  , objpos  = Origin "generated object: step 4b"