+++ ReleaseNotes.md
@@ -2,6 +2,13 @@
 
 ## Unreleased changes
 
+* [Issue #855](https://github.com/AmpersandTarski/Ampersand/issues/855) Minor enhancement in CLASSIFY statement
+* Add semantics to difference in API or INTERFACE keywords.
+
+## v3.11.5 (23 november 2018)
+
+* [Issue #628](https://github.com/AmpersandTarski/Ampersand/issues/628) Fixed a performance issue for specific queries.
+
 ## v3.11.4 (27 october 2018)
 
 * Minor changes in the releaseing of ampersand